refactor: Change unsigned long to uint32_t
Fix #18

+++ unyte_utils.h
@@ -31,9 +31,9 @@ struct unyte_segment
 struct unyte_metadata
 {
   /* Metadatas */
-  uint16_t src_port;            /* Source port */
-  unsigned long src_addr;       /* Source interface IPv4*/
-  unsigned long collector_addr; /* Collector interface IPv4*/
+  uint16_t src_port;       /* Source port */
+  uint32_t src_addr;       /* Source interface IPv4*/
+  uint32_t collector_addr; /* Collector interface IPv4*/
 };
 
 /**
@@ -56,15 +56,15 @@ struct unyte_minimal
   char *buffer;
 
   /* Metadatas */
-  uint16_t src_port;            /* Source port */
-  unsigned long src_addr;       /* Source interface IPv4*/
-  unsigned long collector_addr; /* Collector interface IPv4*/
+  uint16_t src_port;       /* Source port */
+  uint32_t src_addr;       /* Source interface IPv4*/
+  uint32_t collector_addr; /* Collector interface IPv4*/
 };
 
 typedef struct unyte_socket
 {
-  struct sockaddr_in *addr;          /* The socket addr */
-  int *sockfd;                        /* The socket file descriptor */
+  struct sockaddr_in *addr; /* The socket addr */
+  int *sockfd;              /* The socket file descriptor */
 } unytesock_t;
 
 struct unyte_minimal *minimal_parse(char *segment, struct sockaddr_in *source, struct sockaddr_in *collector);
